Updated commands
Updated the commands, removed the PatchAppleBacklight_v2 folder as it
does not work.
The setup.commnad  implements the first part of the fix among other
things and then the Finish.command will complete the fix as it installs
the config.plist with a Patch in KextsToPatch as well as places the
needed .aml file in the appropriate location. It may be necessary to
reboot twice, rebuilding the kext cache each time for the KextsToPatch
to work.